+++ Models/RequestModels/Transactions/VendorInvoiceTxn/VendorInvoiceTxnUpdateRequestModel.cs
@@ -21,12 +21,12 @@
 
         // Tab2
         public List<InvoiceFessDetailList>? invoiceFeeDetails { get; set; }
-        public int? professionalFeeAmt { get; set; }
-        public int? govtOrOfficialFeeAmt { get; set; }
-        public int? otherChargesAmt { get; set; }
-        public int? discountAmt { get; set; }
-        public int? discountCreditNoteAmt { get; set; }
-        public int? totalAmount { get; set; }
+        public decimal? professionalFeeAmt { get; set; }
+        public decimal? govtOrOfficialFeeAmt { get; set; }
+        public decimal? otherChargesAmt { get; set; }
+        public decimal? discountAmt { get; set; }
+        public decimal? discountCreditNoteAmt { get; set; }
+        public decimal? totalAmount { get; set; }
 
         // Tab 3
         public DateTime? paymentDate { get; set; }
@@ -34,12 +34,12 @@
         public string? oWRMNo1 { get; set; }
         public string? oWRMNo2 { get; set; }
         public string? paymentCurrency { get; set; }
-        public int? paymentAmount { get; set; }
+        public decimal? paymentAmount { get; set; }
 
         // Tab 4
         public string? customerPONo { get; set; }
         public DateTime? pODate { get; set; }
-        public int? pOValueInclusiveTaxes { get; set; }
+        public decimal? pOValueInclusiveTaxes { get; set; }
         public string? saleCurrency { get; set; }
         public List<SaleInvoiceDetailList>? salesInvoiceDetails { get; set; }
         public List<PaymentInvoiceDetailList>? paymentFeeDetails { get; set; }
